backend/requirements.txt
Django>=3.0,<4.0
djangorestframework
psycopg2-binary  # Para conexão com PostgreSQL
djangorestframework-simplejwt  # Para autenticação JWT
django-cors-headers  # Para permitir requisições CORS
